pyproject.toml
[project]
name = "hyperag"
version = "0.1.0"
description = "An advanced agentic RAG architecture."
requires-python = ">=3.13"
dependencies = [
    "langgraph",
    "llama-index",
    "chromadb",
    "python-dotenv",
    "llama-index-vector-stores-chroma>=0.4.2",
    "openai>=1.95.1",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-openrouter>=0.3.2",
    "llama-index-llms-openai>=0.4.7",
    "pydantic-settings>=2.10.1",
    "prefect>=3.4.8",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-embeddings-jinaai>=0.4.0",
    "requests>=2.32.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "A", "C4"]

[[tool.mypy.overrides]]
module = ["core.*", "ingest.*"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "mypy>=1.16.1",
    "notebook>=7.4.4",
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]
